Modais adicionados e funcionando.

+++ backend/schemas.py
@@ -1,10 +1,10 @@
 from pydantic import BaseModel
 from datetime import date
 
-# TRANSÇÕES #
+# TRANSACOES #
 class TransacaoCreate(BaseModel):
     descricao: str
-    valor: float
+    valor: int 
     data: date
     id_usuario: int
     id_categoria: int
@@ -15,12 +15,13 @@ class TransacaoResponse(TransacaoCreate):
     class Config:
         from_attributes = True
 
-# Usuário #
 
+# USUARIO #
 class UsuarioCreate(BaseModel):
     nome: str
     email: str
     senha: str
+
 
 class UsuarioResponse(BaseModel):
     id_usuario: int
@@ -30,11 +31,11 @@ class UsuarioResponse(BaseModel):
     class Config:
         from_attributes = True
 
-# METAS FUTURAS #
 
+# METAS FUTURAS #
 class MetaCreate(BaseModel):
     nome: str
-    valor_alvo: float
+    valor_alvo: int
     data_limite: date
     id_usuario: int
